requirements.txt
pycrypto==2.6.1
scapy==2.4.0
pycryptodomex==3.4.5

